refactor: split classic battle helpers
